fix(rebrand): rename gittensory-repo-focus-manifest.ts module
Renamed src/config/gittensory-repo-focus-manifest.ts to
loopover-repo-focus-manifest.ts and updated every import/reference
across test/unit, src/, packages/loopover-engine, and
scripts/check-manifest-drift.mjs. Also renamed its own test file to
match. Completes the full-cutover rename of this module.
