+++ src/config/loopover-repo-focus-manifest.ts
